refactor(Judge): Player 및 Dealer 게임 결과 객체 패키지 변경과 테스트 추가 및 자료구조 변경
- 기존: view 패키지 내에서 HashMap 이용

- 변경: domain으로 패키지 변경 및 EnumMap 사용
