+++ src/test/java/domain/DealerScoreTest.java
@@ -1,0 +1,23 @@
+package domain;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+class DealerScoreTest {
+
+    @Test
+    @DisplayName("PlayerScore가 입력될 경우 딜러의 성적이 플레이어와 정반대로 결정된다")
+    void ComputeDealerScore() {
+        PlayerScore firstPlayerScore = new PlayerScore("roy", GameResult.LOSE);
+        PlayerScore secondPlayerScore = new PlayerScore("hoy", GameResult.LOSE);
+        List<PlayerScore> playerScore = List.of(firstPlayerScore, secondPlayerScore);
+
+        DealerScore dealerScore = new DealerScore(playerScore);
+
+        assertThat(dealerScore.getWin()).isEqualTo(2);
+    }
+}
