+++ src/test/java/domain/PlayerScoreTest.java
@@ -1,0 +1,20 @@
+package domain;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+
+class PlayerScoreTest {
+
+    @Test
+    @DisplayName("플레이어 이름과 게임 결과를 입력하면 PlayerScore가 정상적으로 생성된다.")
+    void createPlayerScore() {
+        String name = "roy";
+        GameResult gameResult = GameResult.WIN;
+
+        PlayerScore playerScore = new PlayerScore(name, gameResult);
+
+        assertDoesNotThrow(() -> playerScore);
+    }
+}
